requirements.txt
google-genai
google-api-python-client
python-dotenv
requests
feedparser
pytest>=7.0
